+++ Changelog.md
@@ -2,6 +2,12 @@
 
 Toutes les modifications notables apportées au projet VBWinDiff sont documentées dans ce fichier.
 
+## [1.14] - 2025-05-17
+### Ajouté
+- Algorithme TextDiffToHtml ajouté, en plus de WinDiff et WinMerge
+- Option Découpage des paragraphes en phrases en conservant la ponctuation
+- Avec la gestion de espaces insécables, remplacé aussi le signe … en ...
+
 ## [1.13] - 2024-08-04
 ### Ajouté
 - Tirets moyens remplacés par tirets courts, avec les espaces insécables